♻️(back) better managed exception with moodle api
When api call are made on the moodle API, some exceptions can be thrown
and are not caught. First the MoodleException can be thrown when the api
return a 200 but the keyword exception is present in the response body.
Also, a netwotrk exception can be thrown and we must catch it.