🐛(backend) zip archive of contracts by organization
When a user has access to several organizations and those
ones are implied in the same `CourseProductRelation`, when
the user asks to download an archive of all contracts, he gets all
contracts from all the organizations he has access to. Before the fix,
only one parameter (CPR or Org) could be used for this endpoint.
Now, the endpoint accepts both parameters  `CourseProductRelation`
UUID and  `Organization` UUID. Finally, only the contracts that
are attached to a specific organization are now archived.

Fix #714