🔧(tray) complete admin nginx conf
Currently, on the back office, if a user tries to go directly on a resource
page, it gets a 404 error. As we made a static export of our next application,
we can't take apart of the fallback option from the getStaticPaths method. In
order to prevent to have to build our next application to run through a node
server, we decide to simply add a nginx location rule to try to find the right
html file for dynamic routes containing an id parameter.

We did this modification for local environment but not for deployment purpose.