🐛(backend) prevent to sign a document when signatories are missing
To avoid the situation where the organization has not set any
owner members, it causes the contract to have only the student
to sign the document where we actually need the signature of both
parties to make the contract legit and allow the student to confirm
is engagement for the upcoming course linked to his order. To prevent
this situation, we've added a method that will just check if the
organization has set at least 1 owner before creating the signature
procedure.

Fix #713