✨(back) endpoint to retrieve a course_run filter on its course_link
To complete the import process from edx, we need to create a new
endpoint able to return a course run filter on its course_link. The
course_link parameter is mandatory and the AdminCourseRunSerializer is
used, it contains all the information we need in the import process.
This endpoint is a server to server endpoint, the HasApiKey permission
is used here.