♻️(back) remove usage of Site model in payment mail
In the mail sent when a payment is successful, we are using the Site
model to get the url and the name. We are now using the Site model for
joanie url, not for the catalog url and name. We now use the settings
JOANIE_CATALOG_BASE_URL and JOANIE_CATALOG_NAME dedicated for this.