♻️(backend) improve contract document title
Currently the document uploaded on the signature provider uses
the contract definition title as title. This is weird as for a course,
all contracts will have the same name. In order to improve that we decide
to build a specific title according to the current date, the related course
and the order primary key.

Resolve #729