🐛(backend) convert validity period in ms for Lex Persona backend
On Lex Persona, each workflow steps have a validity period. We sets this
one according to the settings JOANIE_SIGNATURE_VALIDITY_PERIOD but
this settings is expressed in seconds and it appears Lex Persona expects
a duration in ms.

Furthermore, we decide to update the default
JOANIE_SIGNATURE_VALIDITY_PERIOD to the max value accepted
by Lex Persona which is 99 days.