🐛(backend) enrollment search by query for backoffice django
For the django backoffice, when we use the search bar, there was
an error thrown where it cannot resolve the keyword 'organization'
into the fields. The path to obtains the organization field was not set
up properly in the search fields of the admin class. Now every search
options work as they should be, we can pass an organization title and it
will retrieve the enrollments found matching the required user to set for
filter and the organization title string.

Fix #753