🔇(moodle) remove error logs when user is not found
Each time a user is not found on Moodle, a logger.error message is sent. That
means that for all new users, an event will be sent that is weird. In order to
fix that, we decrease the log level to info. In this way, this log will be bound
 to an error event if needed.