🐛(user) prevent newsletter subscription when user save fails
On user creation, the newsletter subscription was done before the user
was saved. This could lead to a situation where the user was not saved
but the newsletter subscription was. This commit moves the newsletter
subscription to after the user is saved. This way, if the user save
fails, the newsletter subscription will not be done.