✨(backend) new sections for contract definition template
There are two new sections in the course syllabus of Richie that
we need to extract the information through RDF attributes
for our contract_definition context. We've updated the processor
context for the accessibility section, and for the required equiment
it was already in gathered through the 'about' section.

Fix #764