✨(backend) admin certificate view force organization filter
To prevent in django admin to timeout when there will be numerous entries
when calling the certificate view, we force the user to select an
organization first to get some results. The user will be able to use the
search bar to filter out more through the organization's certificates results
with the username, email, or first name of the learner.

Fix #768