🗃️(backend) store Certificate images through DocumentImage model
Currently, we are storing certificate images into localized_context json field
as base64 encoded string. This is weird as it takes a lot of space in database
then we duplicate images in each certificate where they are used.

That's why, we decide to stop that and store Certificate images through a
relation to the DocumentImage table.