🐛(backend) contract context course section as strings
When we generate the context for a contract, we would have
datetime value for course start and course end, and, timedelta
value for course effort, and finally Decimal value for course price.
The contract field context has a DjangoJSONEncoder which will transform
those values with string representation but it would cause some errors
when we render them in the contract template, because those values
are not strings anymore when we use them. Our template tags that
formats our date and duration value were awaiting string values,
and not native python objects such as datetime, or timedelta. To
avoid facing this situation, we now transform those values of
dates and duration into iso 8601 string format before returning
the context and save it to the contract. We want the value to be
prepared at their creation to always have control on it.

Fix #741