✨(backend) submit installment payment when one has failed for order
New endpoint to submit an installment payment when one has failed
on an order. If the `credit_card_id` is passed, then the payment
will go through the `create_one_click_payment`. Else, if no
`credit_card_id` is passed in the payload, the endpoint will
use `create_payment` method. In order to simplify our code on
the ownership of a credit card, we have decided to add a new
Manager class for the `CreditCard` model. The manager class
has one method that retrieves the credit card object if the
queryset matches the input `credit_card.id` and the `owner_id`.