♻️(back) when an installment is paid propagate it in the payment backend
Each installment in a payment schedule now have an id. It is easier to
identify it. When an installment is paid, we can now use it in the
payment backend for all existing payment method (create_payment,
create_one_click_payment and create_zero_click_payment). Doing this, the
`_do_on_payment_success` and `_do_on_payment_failure` methods are able
to manage an installment lifecycle. There is no more need to manage it
in the payment schedule task.