♻️(backend) refactor _get_common_payload_data in payment backend
We have decided to pass a billing address as an Address object
instead of a dictionary representing the address. This change has
led us to modify the endpoint `submit` as well because it was
still passing a billing address dictionary to `create_payment`,
`create_one_click_payment` and `tokenize_card` methods.