🩹(backend) add missing order flow transition sources
Some state source were missing for "complete" and "pending_payment"
transitions. Indeed if the last payment failed, then when it is
resolved, the order transition to "complete" state.
Then, when order has a "no_payment" or "failed_payment" state, when the
installment is resolved, it should transition to "pending_payment"
state.