✨(backend) add `payment_provider` name information to `CreditCard` model
The credit card was not linked to a payment provider which would cause issues
if the payment provider is changed, then, all the existing credit card won't be
usable as they will not be recognized by the new payment provider. We added
a new field called `payment_provider` that stores the name of the payment
provider used to tokenize the credit card for a user. We've also added
the payment provider name on the creation of the credit card depending
the active payment provider.

Fix #814