♻️(backend) remove `has_consent_to_terms` for `Order` model
From now on, the terms and conditions (CGV in french) must be
specific for every organization. This means we can't have a global
one for the whole platform. The terms and conditions will be
injected into the contract's context, so the Order model does need
to track down if the user has accepted those, since he will accept them
while signing the contract.