♻️(backend) remove usage of `SiteConfig` for terms and conditions
The terms and conditions will be written and stored into the
contract definition body field. We don't need to prepare
anymore it for the contract definition context. We've
updated the template html in charge of preparing the
contract by taking away the `terms_and_conditions`
in the `Appendices` section. For now, the usage of
SiteConfig is deprecated to prepare the  terms and
conditions for a contract in markdown.