♻️(back) refactor scheduling generator to match expected behavior
The scheduling generator was not fully implementing what is expected in
the article L 221-19 of the Consumer Code.
The withdrawal date is the good one but once the first payment made the
second one can not occurs before the start of the course run. Then the
others payments can be calculated based on this second payment.