🐛(backend) moodle user creation must raise EnrollmentError exception
Currently, in MoodleBackend,  if during a `set_enrollment`, a user must be
created and this task failed, a MoodleUserCreateException is raised. This
exception is not caught by the method `set` of the Enrollment model and it
should not be manage as this is too specific exception.
But currently, if this error is raised when we try to auto enroll a user during
an order validation, the order validation can be rollback. That's why we have to
 improve this error management and raise an EnrollmentError instead.