♻️(backend) deprecate `has_consent_to_terms` for Order model
From now on, the terms and conditions (CGV in French) must be
specific to each organization. We can no longer use a global
version for the entire platform. These terms will be included
directly in the contract's context, so the Order model no longer
needs to track user acceptance, as this will happen during contract
signing.

Fix #816