🎨(backend) update context for contract for terms and conditions
The terms and conditions will be written and stored in the
contract definition body field. There's no need to prepare them
separately for the contract context anymore. We've updated the
generation of context by removing the key `terms_and_conditions`
in the preparation. In this commit, we have also updated
the contract definition html template for the section "Appendices".
The use of `SiteConfig` to prepare terms and conditions in html
is now deprecated.