🔧(settings) remove payment backend default value
By default, JOANIE_PAYMENT_BACKEND settings was set to
"joanie.payment.backends.payplug.PayplugBackend". We do not want to use this
payment provider now and furthermore it's weird to enable it by default.
About the job tray, we set the payment backend to the dummy one, to be able to
exec some logic which needs a payment backend without to have to provide
further configuration.