✨(backend) allow to configure how the withdrawal data is computed
We have added two new settings variables to be able to
choose the country's calendar depending on where the application
is used geographically. We have added a convenient method
to get the instantiated calendar object. The other setting
allows to set the period of days (as a positive integer)
a consumer has the right to withdraw when signing a contract.

Fix #789