👔(backend) certificate order filter return enrollment degrees
With the data migration, we imported degrees and attestation of achievement. To
prevent to have to create all the data structure (product, product course
relation, orders), for degrees, we simply linked them to their enrollment and we
 used a "degree" certificate template. But there is one drawback to do that, the
 api allows to filter certificates according to their relation to order and
enrollment models. In short, imported degrees are currently returned with
legacy attestation of achievement. In order to fix that, we propose to tweak
the certificate filter to include those specific degrees with order certificates