🗃️(backend) store Order images through DocumentImage model
Currently, we are storing order images (organization logo) into json
field as base64 encoded string. This is weird as it takes a lot of space
in database then we duplicate images in each certificate where they are used.

That's why, we decide to stop that and store Order images in a DocumentImage.