👔(backend) update condition to transition order to pending_payment
Currently, an order in pending state is able to transition to pending_payment
once no installment has been refused. That means we are able to transition to
this state order to which we never try to pay an installment that is weird.
Indeed, only orders with the first installment paid and all others installment
not refused should be allowed to transition to pending_payment state