✨(backend) allow to generate payment schedule for any kind of product
Currently, the payment schedule was generated on the signature submit. But this
was weird as some product can have no contract and in this case, no payment
schedule is generated. So in order to support this kind of product, we move the
payment schedule generation logic on pending transition success.