♻️(back) use settings.AUTH_USER_MODEL and change deletion to cascade
We need to change relation with the user model to delete them when a
user is deleted. For this, the on_delete for the ForeignKey is changed
to CASCADE. To be really used, we also need to stop using the User model
directly but instead use the settings AUTH_USER_MODEL as explain in the
django documentation.