✨(back) create a management command to bulk delete users
In the Mork project, a redis set will be defined with a list of user to
delete using there usename. Joanie is then responsible to do this
deletion in batch. For this, we created a management command, reading
the set in redis and then triggering the user deletion in a celery task.