✨(backend) debit installment if due date is current day
We have added in post transition success to 'pending' state
of an order that will trigger a payment to let the user
have access to the course immediately when the installment's
due date is on the current day. This case may happen when
we generate the payment schedule and if the course has
already started, the 1st installment due date of the order's
payment schedule will be set to the current day. Since we only
debit the next night through a cronjob, we need to be able to
make the user pay to have access to his course, and avoid that the
has to wait the next day to start it.

Fix #913