✨(backend) cast due_date to date, amount to Money in payment_schedule
The `OrderPaymentScheduleDecoder` was returning a string
representation of the `due_date` instead of a `datetime.date` object.
The same way, it was returning a string representing the `amount`
instead of a `Money` object. This behavior complicated
comparisons and operations, as handling both strings
and money and date objects increased the complexity.
To simplify date and money handling and ensure consistency,
a cast was added to the decoder, converting the string
from the database into a `datetime.date` object for `due_date`,
and the `amount` into a `Money` object. This change ensures that
`due_date` is always a `date` object, and `amount` is always
a `Money` object. It makes it easier to work with
throughout the codebase. For our fellow developers, they
can now freely pass strings to prepare their payment schedule
with the `OrderFactory` or the `OrderGeneratorFactory`
in tests, where for the fields `due_date` and `amount`, they
both will be casted to their respective types after being created.