✨(backend) refund state in order flow
We want to be able to refund the installments paid on an order
for a student. This state is only accessible when the order is
cancelled. Some adjustments were made for the OrderGeneratorFactory
to use this new state 'refund' in tests. There will be very few
cases where an order can be refund, for example : the session
is canceled, the professor is on sick leave, or the admin
decides to refund the installments for the student in specific
cases.