✨(backend) payment backend refund and cancel action
For now, we have added the refund and cancel action with
payment backend Lyra. It allows to refund or cancel
a transaction created with the payment provider. The payment
provider decides whether the transaction can be cancelled
or refunded, depending on the capture date of the amount.
For our case, we let the payment provider decide the
operation to do, and we create a transaction and a
credit note when we receive the notification through
our payment webhook. Adjustments were done to
set an installment as refunded once we receive the
notification from the payment provider.