🧪(backend) reproduce timeout encounter with delete signing procedure
We have encountered a couple of time a timeout with the
signature provider. We added a new test to simulate what
happens to the contract object when it encounter a time
on the endpoint `delete_signing_procedure`. This method
is called when we submit for a signature a contract, if the
contract exists then, we verify if the order's contract context
has changed since last submission. If it has changed, we must
delete the old contract at the signature provider and submit
the new one with the updated context. The issue we are facing
is with the timeout, we don't get to submit the new contract
again, and the object that was supposed to be updated still
have the old signing reference backend on his field
signature_backend_reference. And the new contract never
gets submitted. When the organization wants to sign on his
turn, the signature_backend_reference does not exist on the
side of the signature provider. The deletion was made, but not the resubmit.