✨(backend) get signing progress on document for signature backends
Instead of waiting for the webhook that a document has been signed
from the signature provider (the cause would be a latency to send us
the notification event), we have decided to add a new endpoint
on the backend signature provide to get the information if whether
the student or the organization have signed the document. The new
method returns a dictionary with boolean value that informs the
progression of signing procedure of a document.