🥅(backend) improve error catching of signature notification
Currently, when your signature provider sends a notification, if
our internal logic raise an error, an error HTTP Response is
returned. Then the signature provider will retry to send this
notification and it could lead to heavy latency to receive
other notification. So we improve error catching from your
side to prevent to return Error HTTP response if an error
is raised from an internal server error.