🚧(backend) student_signed_on with signature provider on serializer
For the serializer client ContractLightSerializer, whether the
notification from the signature provider takes some time
to let us know that the student has signed, when the viewset is called,
it will trigger our serializer method that will check if the
student has finished signing his part on the document.
That allows the frontend to get the latest information when
the student gets to his dashboard of course orders.