✨(backend) admin api refund and cancel endpoints
For admin users, we have created two new endpoints
to change the state of an order. The first one is
to cancel an order, and once the order is cancelled,
the admin user may use the refund endpoint to trigger
the refund of all paid installments on an order's
payment schedule.