✅(backend) create main invoice when state is not draft OrderFactory
We have added in the post_transition_success of the order flows to
enable a payment for the user if the course has already started, which
will allow the student to enter the course directly instead of waiting
overnight. By adding this, it seems like somes tests failed suddenly.
We've decided to fix the `OrderFactory` class by creating a main
invoice when the state of the Order is not in `draft`. It eases the
life of developers while testing that specific state for an order,
they won't have to create the main invoice by hand in a test after
creating the order.