✨(backend) add transaction and children invoice OrderGeneratorFactory
The `OrderGeneratorFactory` was missing the creation of Transaction
objects in the state where we simulate that an installment or
more were paid. In fact, when we create an order with this factory
for test purposes, we need to have an Invoice (main invoice),
children invoices and a related transactions to reflect the
payments made for an installment in the payment schedule.