⬇️(back) downgrade django-storages to version 1.14.3
In version 1.14.4 there is one modification made that lead to a
breaking change in Joanie as explain in jschneier/django-storages#1430.
We have to wait a newer version with a fix to have the previous behaviour. This
fix is related to a security issue in django. This security is fixed in version
4.2.14, and we already use this version, so we are safe.