👔(backend) update `handle_notification` signature backend logic
Currently, on LexPersona backend, we validate student signature on
recipientFinished event and organization signature on workflowFinished event.
This is weird as when an organization signs, two events are send by LexPersona
(recipientFinished + workflowFinished). As consequence, when the organization
signs, `student_signed_on` is updated on `recipientFinished` event and
`organization_signed_on` is set on `workflowFinished`. So in order to fix that,
we do not use anymore the `workflowFinished` event. Instead, as our workflow is
static, we can assume to set student_signed_on on the first `recipientFinished`
event then set organization_signed_on the second time we receive a
`recipientFinished` event for the related contract.