✨(backend) allow admin user to create update enrollment for user
It was forbid for an admin user to create/update an enrollment
on a closed course run and opened course run. Now it is eligible.
The logic has been put into the serializers on the client API
to preserve the old logic regarding students to prevent them
from enrolling to a closed course run.

Fix #962