✨(backend) add new default template for contract
Currently, we only had one template for contract and it's
branded with the Unicamp logo. We decided to add a contract
without a brand logo for organizations who desire to sell courses
with a contract with their own logo only on the contract. There are
now two available options for the name of the contract definition.
This development involved creating migration files with the new
choice, and we made a rollback function if we need it in the future.
We also made a symbolic link for the stylesheet of the contract
definition unicamp, because for now, it's exactly the same
stylesheet as for the contract definition default stylesheet.

Fix #866