👽️(backoffice) add api interfaces, hooks and types for Skill & Product
Product model has new fields `certification_level`, `skills` and `teachers` and
we want to be able to edit those fields through the product detail view. So as a
first step we add all stuff to be able to query those resources and update
Product types to include those new fields.