🐛(backend) cancel and refund installment on order
With the actual payment provider backend we are using,
it seems like we don't get the events through our notification
URL when the action is 'refund'. It's problematic because
we use the notifications to update our objects. After
couple of attempts to add the ipnTargetUrl into the payload
of the cancel/refund request, it was not possible to receive
the event afterwards. Knowing that, I've decided to
update the objects after the call with the response back
from the payment provider. There might other options
to make sure that the action went successfully (getStatus of
a transaction for e.g)