🐛(backend) cancel and refund installment with lyra backend
With the actual payment provider backend we are using (Lyra),
it seems like we don't get the events through our notification
URL when the action is 'refund'. It's problematic because
we use the notifications to update our objects. After
couple of attempts to add the ipnTargetUrl into the payload
of the cancel/refund request, it was not possible to receive
the event afterwards. Knowing that, we now update our objects
once we receive the response from the call of that endpoint.
Only for cancel or refund situation, we don't rely on the
notification anymore.

Fix #793