👔(backend) update get_transaction_references_to_refund
When we have many installments that are paid in the payment
schedule of an order, the method would not return exactly what was
expected when using only the amount as comparison. Now, it checks
if the transaction reference is present, and also checks if the installment
id is present into the value part of the dictionary. Since we need to
provide more information for the `cancel_or_refund` endpoint,
we added a the installment reference to update our object
when the call to the endpoint is successful.