♻️(backend) dummy payment cancel or refund endpoint
We are now using Lyra as a payment backend. We have
discovered that Lyra does not let us handle through
notifications our refunded installments and forces
us to use the response of that call. To mimick the
behavior, we have decided to change the behavior
as well of the dummy payment backend to update
the objects once we call the method to cancel or
refund an installment for an order.