🧑‍💻(debug) update payment debug view with `OrderFactory`
The `OrderGeneratorFactory` used for the debug payment view
was not appropriate because it creates a transaction when
the state of the order in in pending payment. When we
create a payment with the debug view, and we want to refund it,
the method `get_transaction_references_to_refund` would find
2 different transactions with the  same amount to refund. Only
one out of two is known to the payment provider, and thus,
it would raise an error while attempting to refund the
order from the backoffice of Joanie.