✨(backend) credit card can have multiple owners
We want to able a group of organization members to pay
their orders with a common credit card. We discovered
that we cannot do this action because of the unicity of
the token value received from the payment provider when
tokenizing a credit card. That lead us to block a new entry of
the same credit card for another user. To enable this, we
have decided to track down the different users who will
use the same credit card, and by this, introduce a
ManyToMany relation field named owners to the CreditCard
model. We kept the unicity of token this way, because
we don't want duplicates of the same card in our
database.

Fix #995