✨(backend) deletion of credit card when shared with owners
We can only allow to delete a credit card if there are no more
pending payments on a credit card of an owner on his orders.
When the credit card is shared with other users, the user
that requests to delete card will just be taken away from
the relation, the other users will still have access to the
credit card. Only when the user is alone as an owner on the
credit card that we can allow the deletion of the credit card
object.