✨(backend) add unique constraint on order model
Currently, it was not possible to create a new order
when an order with the triplet owner-course-product
or owner-enrollment-product already existed that were
in the following states : `refunding` to `refunded`.
Only the when the order was in state `cancelled` we
could create a new order. Now, we have decided to add
`refunded` and `refunding` state of orders that are not
taken in account for uniqueness.

Fix #1006