✨(backend) delete orders that are stuck in specific states
We have encountered some issues in the distribution of students
to organization when they enroll. We have discovered that the
orders in states `to_sign` and `signing` were causing that effect.
Those orders exist because students have decided not to complete
the process to sign the contract, and the order would stay in that state.
The issue is that it takes one available seat on the session. We have added
a django command that is responsible to clean up those orders that are
considered as stuck. We also added the deletion of orders for product
type certificate where the student opened the sales tunnel to
purchase the certificate but never entered their payment informations.
To avoid this in the future, we have added a minimum time limit of
tolerance on the latest update of those objects. If they reach this
time limit, we now delete the order and the related objects.

Fix #1018