✨(backend) add start and end date on order groups model
We want to be able to define a start or an end date
on the order group. Both dates can be setted as well.
The enabled property defines if the order group can
accept new orders when it is in relation with a
course product relation. With this work, the field
`is_active` has become a switch for admin user to
enabled or not the group. Then, the property
`is_enabled` computes the value depending on
the activation of the group and the time constraints.

Fix #1029