🔧(backend) allow to configure JOANIE_LMS_BACKENDS through env var
Currently, LMS Backends list is hardcoded in joanie settings. This is weird as
if we want to setup new LMS that means we have to update those settings then
make a release.
Furthermore, it is also weird from an Open Source point of view as those
settings are not overridable easily. So in order to fix that, we now allow to
set all LMS backend configuration through a JSON array string environment
variable.

If you are a Joanie developer, you may have to update our `env.d/common` file to
 remove old LMS env vars and add the new `JOANIE_LMS_BACKENDS`. Take a look at
 `env.d/common.dist` as an example.