✨(backend) add discount admin viewset
Created a viewset to create and list discounts only.
This will allow admin user to create the discount that
can be added in relation to an order group. We have made
on purpose on the admin api that it is not possible to update
or delete a discount. If you want to do such action, you may
use the django admin backoffice to handle them.