✨(backend) django admin discount view
To ease the life of our admin users, we added into the order
group view the possibility to add a discount, or to take it away.
We also added the counter on the discount view to give us the
information on how many times the discount is used through
the order groups. Finally, we updated the view of the column
where it displays the value of the discount to make it more
human-readable.