🔥(backend) remove `generate_certificate` action in django admin
Since we have fixed the generation of certificates into the BO for
orders with product type certificate, we don't want our admin users
to use the django admin anymore for this task. We decided to remove
the action from the admin courses, products and orders views. It is
now impossible to generate certificate through the django admin views.

Fix #1061