✨(backend) assign order with voucher code from batch order
When the batch order is completed and the orders are
generated with voucher codes, the user can use one
of the voucher and assign himself to an order
in state `to_own`. As a reminder, the batch order
generates orders in `to_own` state with a discount
rate of 1. The user will only need to insert his
voucher code to get access to the course, he won't
have to pay because it has been paid through
the batch order. The voucher code is only linked
to one order in state `to_own`.