👔(backend) update organization assignation logic
Currently only orders with inactive states (canceled, draft, assigned) were
ignored in the organization assignation logic. But orders pending for signature
or without payment method  are take in account but most of time this kind of
orders are abandoned. So it has been ask to ignore this kind of order during
assignation. Furthermore, we did not ignore refunded and refunding orders.

Resolve #1081