✅(backend) fix failing test on auto assignation of organization
Fix a failing flaky unit test that used random module when it
prepared the testing data. It was important to fix it because
it is revealing how the auto assignation of an organization is
on an order when the product has no fee. Also, we have
changed the logic behind the assignation method, we needed
to reflect that change into this test.