✨(backend) cancel orders and delete vouchers of canceled batch order
We need to handle the case when a batch order gets canceled. If
the batch order was in state `completed`, we need to cancel each
generated order in state `to_own` and delete the vouchers associated
to them to make sure that they will not be used. Also, if the orders
that were in `to_own` state are taken by their rightful owners with
the voucher codes, we make sure to unenroll them from the course
runs. Otherwise, if the batch order was not yet in `completed` state,
which means that no voucher nor orders were generated, we can just
mark the batch order as canceled.

Fix #1083