✨(backend) django admin batch order viewset
Batch order django admin app allows to create,
edit a batch order. We made it more user-friendly
by allowing to enter a voucher code in create mode
instead of putting the pk. Also to avoid admin user
to fill the field input with a JSON list of firstname
and lastname, we changed the format to only input :
"firstname lastname" per line. We have also added
the action to cancel a batch order.