✅(backend) fix failing test on refund an order
Fix a failing flaky unit test on refund an order
when its state is canceled.
Once a refund is requested an email should be sent
to notify the user.
The content email based on mjml templates and the
footer part use a variable to fill the catalog base url.
None as default, we override it with the override_settings
before function definition to check its value.